requirements.txt
aiogram
langchain-google-genai
langchain-core
langgraph
python-dotenv
pydantic